+++ golem-cli/src/config.rs
@@ -483,39 +483,3 @@ impl From<AuthSecret> for TokenSecret {
         Self { value: value.0 }
     }
 }
-
-#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-pub struct AccountId(pub String);
-
-impl From<String> for AccountId {
-    fn from(id: String) -> Self {
-        Self(id)
-    }
-}
-
-impl From<&str> for AccountId {
-    fn from(value: &str) -> Self {
-        Self(value.into())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ProjectId(pub Uuid);
-
-impl From<Uuid> for ProjectId {
-    fn from(uuid: Uuid) -> Self {
-        ProjectId(uuid)
-    }
-}
-
-impl From<ProjectId> for Uuid {
-    fn from(project_id: ProjectId) -> Self {
-        project_id.0
-    }
-}
-
-impl Display for ProjectId {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
