+++ golem-cli/src/context.rs
@@ -17,12 +17,13 @@ use crate::auth::{Auth, Authentication};
 use crate::command::shared_args::UpdateOrRedeployArgs;
 use crate::command::GolemCliGlobalFlags;
 use crate::command_handler::interactive::InteractiveHandler;
-use crate::config::{AccountId, AuthenticationConfig};
+use crate::config::AuthenticationConfig;
 use crate::config::{ClientConfig, Config, HttpClientConfig, NamedProfile, Profile, ProfileName};
 use crate::error::{ContextInitHintError, HintError, NonSuccessfulExit};
 use crate::log::{log_action, set_log_output, LogColorize, LogOutput, Output};
 use crate::model::app::{AppBuildStep, ApplicationSourceMode};
 use crate::model::app::{ApplicationConfig, BuildProfileName as AppBuildProfileName};
+use crate::model::AccountId;
 use crate::model::{app_raw, Format, ProjectReference};
 use crate::wasm_rpc_stubgen::stub::RustDependencyOverride;
 use anyhow::{anyhow, bail, Context as AnyhowContext};
