+++ golem-cli/src/model/mod.rs
@@ -27,7 +27,7 @@ pub mod wave;
 pub mod worker;
 
 use crate::command::shared_args::{ComponentTemplateName, StreamArgs};
-use crate::config::{AccountId, AuthenticationConfig, ProjectId};
+use crate::config::AuthenticationConfig;
 use crate::config::{NamedProfile, ProfileConfig, ProfileName};
 use crate::log::LogColorize;
 use anyhow::{anyhow, Context};
@@ -874,4 +874,40 @@ impl From<golem_client::model::PluginDefinition> for PluginDefinition {
 pub struct NewInteractiveApp {
     pub app_name: String,
     pub templated_component_names: Vec<(ComponentTemplateName, PackageName)>,
+}
+
+#[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
+pub struct AccountId(pub String);
+
+impl From<String> for AccountId {
+    fn from(id: String) -> Self {
+        Self(id)
+    }
+}
+
+impl From<&str> for AccountId {
+    fn from(value: &str) -> Self {
+        Self(value.into())
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ProjectId(pub Uuid);
+
+impl From<Uuid> for ProjectId {
+    fn from(uuid: Uuid) -> Self {
+        ProjectId(uuid)
+    }
+}
+
+impl From<ProjectId> for Uuid {
+    fn from(project_id: ProjectId) -> Self {
+        project_id.0
+    }
+}
+
+impl Display for ProjectId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
 }
